Add AudioPlayer component
