+++ components/AudioPlayer/AudioPlayer.module.css
@@ -1,0 +1,70 @@
+.expandedContent {
+  margin: 40px 40px;
+  height: calc(100vh - 80px);
+  width: calc(100vw - 80px);
+  background-color: #000808aa;
+  box-shadow: 0px 0px 10px #ffffff29;
+  border-radius: 40px;
+  backdrop-filter: blur(16px);
+  position: fixed;
+  top: 0;
+  padding: 31px 47px;
+}
+.expandedHeader {
+  display: flex;
+  align-items: center;
+}
+.expandedTitle {
+  color: #32a4a7;
+  font-size: 25px;
+  flex: 1;
+}
+.expandedClearIcon {
+  fill: #707070;
+}
+.expandedBody {
+  text-align: center;
+}
+.expandedSongName {
+  font-size: 40px;
+}
+.expandedSonAuthor {
+  color: #32a4a7;
+  font-size: 25px;
+  margin-bottom: calc(100vh - 31%);
+}
+.expandedAudioSection {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 40px;
+  gap: 30px;
+}
+.playButton {
+  background: #32a4a7;
+  border: 0;
+  border-radius: 50%;
+  padding: 0;
+}
+.playIcon {
+  height: 100px;
+  width: 100px;
+  color: #102914;
+}
+.progressBar {
+  height: 14px;
+  margin-top: 2%;
+  border-radius: 5px 0px 0px 5px;
+  background: #19484a;
+}
+.expandedProgressBar {
+  height: 14px;
+  margin-top: 2%;
+  border-radius: 5px 0px 0px 5px;
+  background: #32a4a7;
+  transition: all 0.4s ease-out;
+}
+.volumeContainer {
+  display: flex;
+  justify-content: center;
+}
